refactor: refactor all of customized code
To improve code quality, refactor all of customized code

Closes #88

BREAKING CHANGE:

Delete unused functions
Correct ambiguous variable names
Separate files by object
