+++ berith/brtapi/wallet_args.go
@@ -1,0 +1,14 @@
+package brtapi
+
+import (
+	"github.com/BerithFoundation/berith-chain/common"
+	"github.com/BerithFoundation/berith-chain/common/hexutil"
+)
+
+type WalletTxArgs struct {
+	From     common.Address  `json:"from"`
+	Value    *hexutil.Big    `json:"value"`
+	Gas      *hexutil.Uint64 `json:"gas"`
+	GasPrice *hexutil.Big    `json:"gasPrice"`
+	Nonce    *hexutil.Uint64 `json:"nonce"`
+}
