+++ berith/selection/vote_result.go
@@ -1,0 +1,17 @@
+package selection
+
+import (
+	"github.com/BerithFoundation/berith-chain/common"
+	"math/big"
+)
+
+/*
+[Berith]
+Object that stores voting results for election
+*/
+type VoteResult struct {
+	Score *big.Int `json:"score"`
+	Rank  int      `json:"rank"`
+}
+
+type VoteResults map[common.Address]VoteResult
