+++ core/state_transition.go
@@ -135,7 +135,6 @@ func NewStateTransition(evm *vm.EVM, msg Message, gp *GasPool) *StateTransition 
 // indicates a core error meaning that the message would always fail for that particular
 // state and would never be accepted within a block.
 func ApplyMessage(evm *vm.EVM, msg Message, gp *GasPool) ([]byte, uint64, bool, error) {
-	//return NewStateTransition(evm, msg, gp).TransitionDb()
 	return NewStateTransition(evm, msg, gp).TransitionDb(msg.Base(), msg.Target())
 }
 
@@ -187,7 +186,6 @@ func (st *StateTransition) preCheck() error {
 // TransitionDb will transition the state by applying the current message and
 // returning the result including the used gas. It returns an error if failed.
 // An error indicates a consensus issue.
-//func (st *StateTransition) TransitionDb() (ret []byte, usedGas uint64, failed bool, err error) {
 func (st *StateTransition) TransitionDb(base types.JobWallet, target types.JobWallet) (ret []byte, usedGas uint64, failed bool, err error) {
 	if err = st.preCheck(); err != nil {
 		return
@@ -226,8 +224,6 @@ func (st *StateTransition) TransitionDb(base types.JobWallet, target types.JobWa
 	} else {
 		// Increment the nonce for the next transaction
 		st.state.SetNonce(msg.From(), st.state.GetNonce(sender.Address())+1)
-		//ret, st.gas, vmerr = evm.Call(sender, st.to(), st.data, st.gas, st.value)
-
 		// [BERITH] staking value false
 		ret, st.gas, vmerr = evm.Call(sender, st.to(), st.data, st.gas, st.value, base, target)
 	}
