+++ core/vm/instructions.go
@@ -764,7 +764,6 @@ func opCall(pc *uint64, interpreter *EVMInterpreter, contract *Contract, memory 
 	if value.Sign() != 0 {
 		gas += params.CallStipend
 	}
-	//ret, returnGas, err := interpreter.evm.Call(contract, toAddr, args, gas, value)
 	//[BERITH]
 	ret, returnGas, err := interpreter.evm.Call(contract, toAddr, args, gas, value, types.Main, types.Main)
 	if err != nil {
