+++ interfaces.go
@@ -119,8 +119,8 @@ type CallMsg struct {
 	GasPrice *big.Int        // wei <-> gas exchange ratio
 	Value    *big.Int        // amount of wei sent along with the call
 	Data     []byte          // input data, usually an ABI-encoded contract method invocation
-	Base     types.JobWallet
-	Target   types.JobWallet
+	Base     types.JobWallet // Sender's wallet type
+	Target   types.JobWallet // Receiver's wallet type
 }
 
 // A ContractCaller provides contract calls, essentially transactions that are executed by
